requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.75.1
grpcio-tools==1.75.1
idna==3.11
importlib_metadata==8.7.1
multidict==6.7.0
numpy==2.3.3
opentelemetry-api==1.39.1
opentelemetry-distro==0.60b1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-grpc==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
propcache==0.4.1
protobuf==6.32.1
requests==2.32.5
setuptools==80.9.0
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
